reorganization fixed: https://github.com/joelnet/MojiScript/issues/110
