+++ examples/conditionals/cond2.mjs
@@ -1,7 +1,7 @@
 import log from 'mojiscript/console/log'
-import cond from 'mojiscript/core/cond'
 import pipe from 'mojiscript/core/pipe'
 import run from 'mojiscript/core/run'
+import cond from 'mojiscript/logic/cond'
 import $ from 'mojiscript/string/template'
 
 const dependencies = {
@@ -13,7 +13,7 @@ const state = 100
 const getTempInfo = cond ([
   [ 0, 'water freezes at 0°C' ],
   [ 100, 'water boils at 100°C' ],
-  [ true, $`nothing special happens at ${0}°C` ]
+  [ () => true, $`nothing special happens at ${0}°C` ]
 ])
 
 const main = ({ log }) => pipe ([
