+++ examples/star-wars-console/main.mjs
@@ -1,8 +1,8 @@
 import W from 'mojiscript/combinators/W'
-import ifElse from 'mojiscript/core/ifElse'
 import pipe from 'mojiscript/core/pipe'
 import join from 'mojiscript/list/join'
 import map from 'mojiscript/list/map'
+import ifElse from 'mojiscript/logic/ifElse'
 import prepend from 'mojiscript/string/prepend'
 import $ from 'mojiscript/string/template'
 import isEmpty from 'ramda/src/isEmpty'
