+++ logic/__tests__/cond.test.js
