+++ logic/__tests__/when.test.js
