+++ logic/cond.js
