+++ logic/when.js
