+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mojiscript",
-  "version": "0.2.4",
+  "version": "0.3.0",
   "description": "MojiScript is an Async First, opinionated, and functional language designed to have 100% compatibility with JavaScript engines.",
   "repository": {
     "type": "git",
